Restructure Fortran model to prepare for `FortranOp`/`MetaOp` distinction
