--- src/Language/Fortran/TypeModel.hs
@@ -1,9 +1,0 @@
-{-# OPTIONS_GHC -Wall #-}
-
-module Language.Fortran.TypeModel
-  ( module Types
-  , module Operator
-  ) where
-
-import Language.Fortran.TypeModel.Types as Types
-import Language.Fortran.TypeModel.Operator as Operator
